+++ actix-files/src/files.rs
@@ -35,9 +35,19 @@ use crate::{
 /// let app = App::new()
 ///     .service(Files::new("/static", "."));
 /// ```
+///
+/// # Multiple Directories Example
+/// ```
+/// use actix_web::App;
+/// use actix_files::Files;
+///
+/// // Serve files from multiple directories with priority order
+/// let app = App::new()
+///     .service(Files::new_from_array("/static", &["./dist", "./public"]));
+/// ```
 pub struct Files {
     mount_path: String,
-    directory: PathBuf,
+    directories: Vec<PathBuf>,
     index: Option<String>,
     show_index: bool,
     redirect_to_slash: bool,
@@ -63,7 +73,7 @@ impl fmt::Debug for Files {
 impl Clone for Files {
     fn clone(&self) -> Self {
         Self {
-            directory: self.directory.clone(),
+            directories: self.directories.clone(),
             index: self.index.clone(),
             show_index: self.show_index,
             redirect_to_slash: self.redirect_to_slash,
@@ -105,19 +115,51 @@ impl Files {
     /// The number of running threads is adjusted over time as needed, up to a maximum of 512 times
     /// the number of server [workers](actix_web::HttpServer::workers), by default.
     pub fn new<T: Into<PathBuf>>(mount_path: &str, serve_from: T) -> Files {
-        let orig_dir = serve_from.into();
-        let dir = match orig_dir.canonicalize() {
-            Ok(canon_dir) => canon_dir,
-            Err(_) => {
-                log::error!("Specified path is not a directory: {:?}", orig_dir);
-                // Preserve original path so requests don't fall back to CWD.
-                orig_dir
-            }
-        };
+        Self::new_multiple(mount_path, std::iter::once(serve_from))
+    }
+
+    /// Create new `Files` instance for specified base directories from an array.
+    ///
+    /// # Argument Order
+    /// The first argument (`mount_path`) is the root URL at which the static files are served.
+    /// For example, `/assets` will serve files at `example.com/assets/...`.
+    ///
+    /// The second argument (`serve_from`) is an array of paths on disk at which files are loaded.
+    /// For example, `["./dist", "./public"]` would serve files from both directories in order of priority.
+    pub fn new_from_array<T: Into<PathBuf> + Clone>(mount_path: &str, serve_from: &[T]) -> Files {
+        Self::new_multiple(mount_path, serve_from.iter().cloned())
+    }
+
+    /// Create new `Files` instance for specified base directories from an iterator.
+    ///
+    /// # Argument Order
+    /// The first argument (`mount_path`) is the root URL at which the static files are served.
+    /// For example, `/assets` will serve files at `example.com/assets/...`.
+    ///
+    /// The second argument (`serve_from`) is an iterator of paths on disk at which files are loaded.
+    /// For example, `vec!["./dist", "./public"].into_iter()` would serve files from both directories in order of priority.
+    pub fn new_multiple<I, T>(mount_path: &str, serve_from: I) -> Files
+    where
+        I: IntoIterator<Item = T>,
+        T: Into<PathBuf>,
+    {
+        let directories = serve_from
+            .into_iter()
+            .map(|path| {
+                let orig_dir = path.into();
+                match orig_dir.canonicalize() {
+                    Ok(canon_dir) => canon_dir,
+                    Err(_) => {
+                        log::error!("Specified path is not a directory: {:?}", orig_dir);
+                        orig_dir
+                    }
+                }
+            })
+            .collect();
 
         Files {
             mount_path: mount_path.trim_end_matches('/').to_owned(),
-            directory: dir,
+            directories,
             index: None,
             show_index: false,
             redirect_to_slash: false,
@@ -403,7 +445,7 @@ impl ServiceFactory<ServiceRequest> for Files {
 
     fn new_service(&self, _: ()) -> Self::Future {
         let mut inner = FilesServiceInner {
-            directory: self.directory.clone(),
+            directories: self.directories.clone(),
             index: self.index.clone(),
             show_index: self.show_index,
             redirect_to_slash: self.redirect_to_slash,